+++ docs/releasenotes/releasenotes.rst
@@ -8,6 +8,11 @@ For Dashticz's **master** version Release Notes go to: https://dashticz.readthed
 Recent changes
 ---------------
 
+Fixes
+~~~~~~
+
+* Standby: Prevent click to activate a Dashticz block
+
 v3.8.1 Beta (14-4-2021)
 -----------------------